refactor: moved loadMoreMessages to useMessages
